Updated AvatarGroup
